BASINS40 *.vbproj: added reference to MapWinInterfaces.dll to satisfy SharpDevelop
atcSWMMConduits: corrected Catchment to Conduit

atcSynopticAnalysis/frmSynoptic.SaveSettings: use GetSetting to check whether we need to delete setting instead of relying on Try/Catch

atcUtility/modFile.clsLinesInFile.NextLine: use StringBuilder instead of &= for speed

+++ BASINS40/atcSynopticAnalysis/frmSynoptic.vb
@@ -132,10 +132,9 @@ Public Class frmSynoptic
                 If mnuReverseGroupOrder.Checked Then
                     SaveSetting("Synoptic", "Defaults", "ReverseGroupOrder", "True")
                 Else
-                    Try
+                    If GetSetting("Synoptic", "Defaults", "ReverseGroupOrder").Length > 0 Then
                         DeleteSetting("Synoptic", "Defaults", "ReverseGroupOrder")
-                    Catch 'ignore error if setting is already not there
-                    End Try
+                    End If
                 End If
             Catch 'Saving setting is not critical, ignore errors
             End Try
